fix: edit 15 PW

+++ controllers/cards.controllers.js
@@ -1,13 +1,14 @@
 const CastError = require('mongoose/lib/error/cast');
 const ValidationError = require('mongoose/lib/error/validation');
 const Card = require('../models/card');
-const { CARD_NOT_FOUND, CREATED_CODE } = require('../utils/constants');
+const { CARD_NOT_FOUND, CREATED_CODE, NO_RIGHTS } = require('../utils/constants');
 const BadRequestError = require('../utils/errors/BadRequestError');
 const NotFoundError = require('../utils/errors/NotFoundError');
+const ForbiddenError = require('../utils/errors/ForbiddenError');
 
 module.exports.createCard = async (req, res, next) => {
   try {
-    let card = await Card.create({ ...req.body, owner: req.user.id });
+    let card = await Card.create({ ...req.body, owner: req.user._id });
     await card.save();
     card = await card.populate('owner likes');
     res.status(CREATED_CODE).send({ data: card });
@@ -37,6 +38,7 @@ module.exports.deleteCard = async (req, res, next) => {
   try {
     const card = await Card.findById(req.params.cardId).populate('owner likes');
     if (!card) throw new NotFoundError();
+    if (card.owner.id !== req.user._id) throw new ForbiddenError(NO_RIGHTS);
     await Card.findByIdAndRemove(req.params.cardId);
     res.send({ data: card });
   } catch (err) {
